+++ packages/web-components/fast-foundation/src/tabs/tab-panel/tab-panel.template.ts
@@ -1,6 +1,9 @@
 import { html } from "@microsoft/fast-element";
 import { TabPanel } from "./tab-panel";
-
+/**
+ * The template for the {@link @microsoft/fast-foundation#TabPanel} component.
+ * @public
+ */
 export const TabPanelTemplate = html<TabPanel>`
     <template slot="tabpanel" role="tabpanel">
         <slot></slot>