+++ sites/fast-website/.eslintrc.js
@@ -1,0 +1,13 @@
+module.exports = {
+    extends: ["@microsoft/eslint-config-fast-dna", "prettier"],
+    rules: {
+        "@typescript-eslint/no-unused-vars": "off",
+        "@typescript-eslint/no-non-null-assertion": "off",
+        "@typescript-eslint/typedef": [
+            "error",
+            {
+                arrowParameter: "false",
+            },
+        ],
+    },
+};