+++ sites/fast-website/src/app/style/constants.ts
@@ -1,0 +1,8 @@
+export const siteBreakpoints: string[] = [
+    "0",
+    "700px",
+    "900px",
+    "999px",
+    "1000px",
+    "1200px",
+];