+++ sites/fast-website/webpack.prod.js
@@ -1,0 +1,23 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+/* eslint-disable no-undef */
+const merge = require("webpack-merge");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
+const TerserJSPlugin = require("terser-webpack-plugin");
+const baseConfig = require("./webpack.common");
+
+module.exports = merge(baseConfig, {
+    mode: "production",
+    optimization: {
+        minimizer: [new TerserJSPlugin({}), new OptimizeCSSAssetsPlugin({})],
+    },
+    output: {
+        filename: "bundle/[name].[hash].js",
+    },
+    plugins: [
+        new MiniCssExtractPlugin({
+            filename: "bundle/[name].[hash].css",
+            chunkFilename: "bundle/[id].[hash].css",
+        }),
+    ],
+});