+++ sites/site-utilities/src/definitions/fast-components/fast-menu-item.definition.ts
@@ -48,6 +48,14 @@ export const fastMenuItemDefinition: WebComponentDefinition = {
                     name: "",
                     description: "The default slot",
                 },
+                {
+                    name: "start",
+                    description: "The start slot",
+                },
+                {
+                    name: "end",
+                    description: "The end slot",
+                },
             ],
         },
     ],