+++ packages/react/fast-components-class-name-contracts-msft/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.8](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components-class-name-contracts-msft@4.8.7...@microsoft/fast-components-class-name-contracts-msft@4.8.8) (2020-06-15)
+
+**Note:** Version bump only for package @microsoft/fast-components-class-name-contracts-msft
+
+
+
+
+
 ## [4.8.6](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components-class-name-contracts-msft@4.8.5...@microsoft/fast-components-class-name-contracts-msft@4.8.6) (2020-06-05)
 
 **Note:** Version bump only for package @microsoft/fast-components-class-name-contracts-msft