+++ packages/react/fast-storybook-design-system-addon/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.38](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-storybook-design-system-addon@1.2.36...@microsoft/fast-storybook-design-system-addon@1.2.38) (2020-06-15)
+
+**Note:** Version bump only for package @microsoft/fast-storybook-design-system-addon
+
+
+
+
+
 ## [1.2.36](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-storybook-design-system-addon@1.2.35...@microsoft/fast-storybook-design-system-addon@1.2.36) (2020-06-05)
 
 **Note:** Version bump only for package @microsoft/fast-storybook-design-system-addon